+++ src/client/java/com/github/theredbrain/foodoverhaul/FoodOverhaulClient.java
@@ -1,6 +1,6 @@
 package com.github.theredbrain.foodoverhaul;
 
-import com.github.theredbrain.foodoverhaul.registry.EntityRegistry;
+import com.github.theredbrain.foodoverhaul.registry.FoodOverhaulEntities;
 import com.github.theredbrain.foodoverhaul.render.block.entity.FoodDisplayBlockEntityRenderer;
 import net.fabricmc.api.ClientModInitializer;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderers;
@@ -12,6 +12,6 @@ public class FoodOverhaulClient implements ClientModInitializer {
 	}
 
 	private void registerBlockEntityRenderer() {
-		BlockEntityRenderers.register(EntityRegistry.FOOD_DISPLAY_BLOCK_ENTITY, FoodDisplayBlockEntityRenderer::new);
+		BlockEntityRenderers.register(FoodOverhaulEntities.FOOD_DISPLAY_BLOCK_ENTITY, FoodDisplayBlockEntityRenderer::new);
 	}
 }