+++ src/main/java/com/github/theredbrain/foodoverhaul/registry/FoodOverhaulCreativeModeTabs.java
@@ -4,14 +4,12 @@ import com.github.theredbrain.foodoverhaul.FoodOverhaul;
 import net.fabricmc.fabric.api.creativetab.v1.FabricCreativeModeTab;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 
-public class CreativeModeTabRegistry {
+public class FoodOverhaulCreativeModeTabs {
 
 	public static CreativeModeTab CREATIVE_MODE_TAB;